+++ zarrs/src/array/array_ops/async_array_read_ops.rs
@@ -84,7 +84,7 @@ pub trait AsyncArrayReadOps: ArrayOps {
     /// The chunks are in order of the chunk indices returned by `chunks.indices().into_iter()`.
     ///
     /// # Errors
-    /// Returns an [`ArrayError`] if the chunk key cannot be encoded or there is an underlying store error.
+    /// Returns an [`ArrayError`] if `chunks` is out-of-bounds of the chunk grid or has an incompatible dimensionality, a chunk key cannot be encoded, or there is an underlying store error.
     #[allow(clippy::missing_errors_doc)]
     async fn async_retrieve_encoded_chunks(
         &self,