+++ zarrs/src/array/chunk_grid/regular_bounded.rs
@@ -369,7 +369,6 @@ mod tests {
         );
     }
 
-    #[allow(clippy::single_range_in_vec_init)]
     #[test]
     fn chunk_grid_regular_bounded() {
         let array_shape: ArrayShape = vec![5, 7, 52];